+++ datanymizer_engine/src/settings/table.rs
@@ -39,3 +39,53 @@ impl Table {
         transform_list
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    mod transform_list {
+        use super::*;
+
+        fn rule_names(t: &Table) -> Vec<String> {
+            t.transform_list()
+                .iter()
+                .map(|(name, _)| name.to_string())
+                .collect()
+        }
+
+        #[test]
+        fn order() {
+            let config = r#"
+                name: table1
+                rule_order:
+                  - first_name
+                  - greeting
+                  - options
+                rules:
+                  options:
+                    template:
+                      format: "{greeting: \"{{ final.greeting }}\"}"
+                  greeting:
+                    template:
+                      format: "dear {{ final.first_name }} {{ final.middle_name }} {{ final.last_name }}"
+                  first_name:
+                    first_name: {}
+                  middle_name:
+                    last_name: {}
+                  last_name:
+                    last_name: {}
+                "#;
+            let t: Table = serde_yaml::from_str(config).unwrap();
+
+            let names = rule_names(&t);
+
+            assert_eq!(names.len(), 5);
+            assert!(names.contains(&"middle_name".to_string()));
+            assert!(names.contains(&"last_name".to_string()));
+            assert_eq!(names[2], "first_name");
+            assert_eq!(names[3], "greeting");
+            assert_eq!(names[4], "options");
+        }
+    }
+}